fixed installation UnicodeDecodeError, thanks to @ezyatev.
You should also use io.open() for opening files instead of the built-in open() function as the io module is consistent from Python 2 to 3 while the built-in open() function is not (in Python 3 it’s actually io.open()). Do not bother with the outdated practice of using codecs.open() as that’s only necessary for keeping compatibility with Python 2.5.
Source: https://docs.python.org/3/howto/pyporting.html#text-versus-binary-data